+ Relatorio de Retorno  -  Relatorio de Financeiro/Retorno

+++ AASPA.Models/Response/MotivoNaoAverbacaoResponse.cs
@@ -8,7 +8,7 @@ namespace AASPA.Models.Response
 {
     public class MotivoNaoAverbacaoResponse
     {
-        public int TotalPorcentagem { get; set; }
+        public double TotalPorcentagem { get; set; }
         public string CodigoErro { get; set; }
         public int TotalPorCodigoErro { get; set; }
         public string DescricaoErro { get; set; }
